requirements.txt
fastapi
uvicorn
python-dotenv
aiohttp
azure-cosmos
passlib
python-jose[cryptography]
bcrypt
python-multipart